lead-capture-ui/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Flask-Login==0.6.3
Werkzeug==3.0.1